% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valmyndigheten.R
\name{get_p}
\alias{get_p}
\title{Getting The election result data from polling station count excel file}
\usage{
get_p()
}
\value{
The cleaned data from excel file
}
\description{
Getting The election result data from polling station count excel file
}
