% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valmyndigheten.R
\name{get_p_a}
\alias{get_p_a}
\title{Getting The election result data including assembly district}
\usage{
get_p_a()
}
\value{
The cleaned data from excel file
}
\description{
Getting The election result data including assembly district
}
