% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valmyndigheten.R
\name{Valmyndigheten_api}
\alias{Valmyndigheten_api}
\title{Class For Converting Raw Data to Parsed Data From Valmyndigheten XML.}
\usage{
Valmyndigheten_api(path)
}
\arguments{
\item{path}{specifies the path for excel file to be downloaded}
}
\value{
Retrieve Parsed data Contents of the Excel for the given path(url)
}
\description{
Class For Converting Raw Data to Parsed Data From Valmyndigheten XML.
}
